Move mercator coveringTiles into a separate file
