+++ src/geo/projection/mercator_transform.ts
@@ -2,10 +2,9 @@ import {LngLat, LngLatLike} from '../lng_lat';
 import {MercatorCoordinate, mercatorXfromLng, mercatorYfromLat, mercatorZfromAltitude} from '../mercator_coordinate';
 import Point from '@mapbox/point-geometry';
 import {wrap, clamp, createIdentityMat4f64, createMat4f64} from '../../util/util';
-import {mat2, mat4, vec2, vec3, vec4} from 'gl-matrix';
+import {mat2, mat4, vec3, vec4} from 'gl-matrix';
 import {UnwrappedTileID, OverscaledTileID, CanonicalTileID, calculateTileKey} from '../../source/tile_id';
 import {Terrain} from '../../render/terrain';
-import {Aabb, Frustum, IntersectionResult} from '../../util/primitives';
 import {interpolates} from '@maplibre/maplibre-gl-style-spec';
 import {PointProjection, xyTransformMat4} from '../../symbol/projection';
 import {LngLatBounds} from '../lng_lat_bounds';
@@ -15,6 +14,7 @@ import {mercatorCoordinateToLocation, getBasicProjectionData, getMercatorHorizon
 import {EXTENT} from '../../data/extent';
 import type {ProjectionData} from './projection_data';
 import {scaleZoom, TransformHelper, zoomScale} from '../transform_helper';
+import {mercatorCoveringTiles} from './mercator_covering_tiles';
 
 export class MercatorTransform implements ITransform {
     private _helper: TransformHelper;
@@ -257,126 +257,7 @@ export class MercatorTransform implements ITransform {
     }
 
     coveringTiles(options: CoveringTilesOptions): Array<OverscaledTileID> {
-        let z = this.coveringZoomLevel(options);
-        const actualZ = z;
-
-        if (options.minzoom !== undefined && z < options.minzoom) return [];
-        if (options.maxzoom !== undefined && z > options.maxzoom) z = options.maxzoom;
-
-        const cameraCoord = this.screenPointToMercatorCoordinate(this.getCameraPoint());
-        const centerCoord = MercatorCoordinate.fromLngLat(this.center);
-        const numTiles = Math.pow(2, z);
-        const cameraPoint = [numTiles * cameraCoord.x, numTiles * cameraCoord.y, 0];
-        const centerPoint = [numTiles * centerCoord.x, numTiles * centerCoord.y, 0];
-        const cameraFrustum = Frustum.fromInvProjectionMatrix(this._invViewProjMatrix, this.worldSize, z);
-
-        // No change of LOD behavior for pitch lower than 60 and when there is no top padding: return only tile ids from the requested zoom level
-        let minZoom = options.minzoom || 0;
-        // Use 0.1 as an epsilon to avoid for explicit == 0.0 floating point checks
-        if (!options.terrain && this.pitch <= 60.0 && this._helper._edgeInsets.top < 0.1)
-            minZoom = z;
-
-        // There should always be a certain number of maximum zoom level tiles surrounding the center location in 2D or in front of the camera in 3D
-        const radiusOfMaxLvlLodInTiles = options.terrain ? 2 / Math.min(this.tileSize, options.tileSize) * this.tileSize : 3;
-
-        const newRootTile = (wrap: number): any => {
-            return {
-                aabb: new Aabb([wrap * numTiles, 0, 0], [(wrap + 1) * numTiles, numTiles, 0]),
-                zoom: 0,
-                x: 0,
-                y: 0,
-                wrap,
-                fullyVisible: false
-            };
-        };
-
-        // Do a depth-first traversal to find visible tiles and proper levels of detail
-        const stack: Array<{
-            aabb: Aabb;
-            zoom: number;
-            x: number;
-            y: number;
-            wrap: number;
-            fullyVisible: boolean;
-        }> = [];
-        const result: Array<{
-            tileID: OverscaledTileID;
-            distanceSq: number;
-            tileDistanceToCamera: number;
-        }> = [];
-        const maxZoom = z;
-        const overscaledZ = options.reparseOverscaled ? actualZ : z;
-
-        if (this._helper._renderWorldCopies) {
-            // Render copy of the globe thrice on both sides
-            for (let i = 1; i <= 3; i++) {
-                stack.push(newRootTile(-i));
-                stack.push(newRootTile(i));
-            }
-        }
-
-        stack.push(newRootTile(0));
-
-        while (stack.length > 0) {
-            const it = stack.pop();
-            const x = it.x;
-            const y = it.y;
-            let fullyVisible = it.fullyVisible;
-
-            // Visibility of a tile is not required if any of its ancestor if fully inside the frustum
-            if (!fullyVisible) {
-                const intersectResult = it.aabb.intersectsFrustum(cameraFrustum);
-
-                if (intersectResult === IntersectionResult.None)
-                    continue;
-
-                fullyVisible = intersectResult === IntersectionResult.Full;
-            }
-
-            const refPoint = options.terrain ? cameraPoint : centerPoint;
-            const distanceX = it.aabb.distanceX(refPoint);
-            const distanceY = it.aabb.distanceY(refPoint);
-            const longestDim = Math.max(Math.abs(distanceX), Math.abs(distanceY));
-
-            // We're using distance based heuristics to determine if a tile should be split into quadrants or not.
-            // radiusOfMaxLvlLodInTiles defines that there's always a certain number of maxLevel tiles next to the map center.
-            // Using the fact that a parent node in quadtree is twice the size of its children (per dimension)
-            // we can define distance thresholds for each relative level:
-            // f(k) = offset + 2 + 4 + 8 + 16 + ... + 2^k. This is the same as "offset+2^(k+1)-2"
-            const distToSplit = radiusOfMaxLvlLodInTiles + (1 << (maxZoom - it.zoom)) - 2;
-
-            // Have we reached the target depth or is the tile too far away to be any split further?
-            if (it.zoom === maxZoom || (longestDim > distToSplit && it.zoom >= minZoom)) {
-                const dz = maxZoom - it.zoom, dx = cameraPoint[0] - 0.5 - (x << dz), dy = cameraPoint[1] - 0.5 - (y << dz);
-                result.push({
-                    tileID: new OverscaledTileID(it.zoom === maxZoom ? overscaledZ : it.zoom, it.wrap, it.zoom, x, y),
-                    distanceSq: vec2.sqrLen([centerPoint[0] - 0.5 - x, centerPoint[1] - 0.5 - y]),
-                    // this variable is currently not used, but may be important to reduce the amount of loaded tiles
-                    tileDistanceToCamera: Math.sqrt(dx * dx + dy * dy)
-                });
-                continue;
-            }
-
-            for (let i = 0; i < 4; i++) {
-                const childX = (x << 1) + (i % 2);
-                const childY = (y << 1) + (i >> 1);
-                const childZ = it.zoom + 1;
-                let quadrant = it.aabb.quadrant(i);
-                if (options.terrain) {
-                    const tileID = new OverscaledTileID(childZ, it.wrap, childZ, childX, childY);
-                    const minMax = options.terrain.getMinMaxElevation(tileID);
-                    const minElevation = minMax.minElevation ?? this.elevation;
-                    const maxElevation = minMax.maxElevation ?? this.elevation;
-                    quadrant = new Aabb(
-                        [quadrant.min[0], quadrant.min[1], minElevation] as vec3,
-                        [quadrant.max[0], quadrant.max[1], maxElevation] as vec3
-                    );
-                }
-                stack.push({aabb: quadrant, zoom: childZ, x: childX, y: childY, wrap: it.wrap, fullyVisible});
-            }
-        }
-
-        return result.sort((a, b) => a.distanceSq - b.distanceSq).map(a => a.tileID);
+        return mercatorCoveringTiles(this, options, this._invViewProjMatrix);
     }
 
     recalculateZoom(terrain: Terrain): void {
